pyproject.toml
[project]
name = "advanced-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
